test: make tests successfully run from forks (#2203)
* use monkeypatch in Cohere unit tests

* fix Langfuse

* nvidia

* fix

* skip AWS authentication when PRs are from dependabot
